updated cl for NVIDIA

+++ src/projcl/kernels/pj_msfn.h
@@ -2,9 +2,9 @@
 #define PJ_MSFN_H__
 
 /* determine constant small m */
-double
-pj_msfn(double sinphi, double cosphi, double es) {
-	return (cosphi / sqrt (1. - es * sinphi * sinphi));
+float
+pj_msfn(float sinphi, float cosphi, float es) {
+    return (cosphi / sqrt (1. - es * sinphi * sinphi));
 }
 
 #endif /* PJ_MSFN_H__ */
